+++ libvast/include/vast/system/actors.hpp
@@ -176,7 +176,7 @@ using accountant_actor = typed_actor_fwd<
   // Record time metric.
   auto(atom::metrics, std::string, time, metrics_metadata)->caf::result<void>,
   // Record integer metric.
-  auto(atom::metrics, std::string, integer, metrics_metadata)->caf::result<void>,
+  auto(atom::metrics, std::string, int64_t, metrics_metadata)->caf::result<void>,
   // Record count metric.
   auto(atom::metrics, std::string, uint64_t, metrics_metadata)->caf::result<void>,
   // Record real metric.
