+++ libvast/test/parse_data.cpp
@@ -35,9 +35,9 @@ TEST(data) {
   CHECK_EQUAL(to_data("true"), data{true});
   CHECK_EQUAL(to_data("false"), data{false});
   MESSAGE("int");
-  CHECK_EQUAL(to_data("+42"), integer{42});
-  CHECK_EQUAL(to_data("-42"), integer{-42});
-  CHECK_EQUAL(to_data("-42k"), integer{-42'000});
+  CHECK_EQUAL(to_data("+42"), int64_t{42});
+  CHECK_EQUAL(to_data("-42"), int64_t{-42});
+  CHECK_EQUAL(to_data("-42k"), int64_t{-42'000});
   MESSAGE("count");
   CHECK_EQUAL(to_data("42"), data{42u});
   CHECK_EQUAL(to_data("42M"), data{42'000'000u});
@@ -57,9 +57,9 @@ TEST(data) {
   MESSAGE("map");
   CHECK_EQUAL(to_data("{}"), map{});
   CHECK_EQUAL(to_data("{+1->true,+2->false}"),
-              (map{{integer{1}, true}, {integer{2}, false}}));
+              (map{{int64_t{1}, true}, {int64_t{2}, false}}));
   CHECK_EQUAL(to_data("{-1 -> true, -2 -> false}"),
-              (map{{integer{-1}, true}, {integer{-2}, false}}));
+              (map{{int64_t{-1}, true}, {int64_t{-2}, false}}));
   MESSAGE("record - named fields");
   CHECK_EQUAL(to_data("<>"), record{});
   CHECK_EQUAL(to_data("<foo: 1>"), (record{{"foo", 1u}}));
