+++ libvast/test/yaml.cpp
@@ -30,7 +30,7 @@ struct fixture {
   fixture() {
     // clang-format off
     rec = record{
-      {"foo", integer{-42}},
+      {"foo", int64_t{-42}},
       {"bar", 3.14},
       {"baz", list{"a", caf::none, true}},
       {"qux", record{
